webhook-orchestrator/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Task Queue
celery[redis]==5.3.4
redis==5.0.1

# HTTP and Security
httpx==0.25.2
cryptography==41.0.8
pyjwt==2.8.0
python-multipart==0.0.6

# Monitoring and Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-jaeger==1.21.0
structlog==23.2.0

# Circuit Breaker and Resilience
circuitbreaker==1.4.0
tenacity==8.2.3

# Validation and Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# GitHub Integration
pygithub==2.1.1

# Codegen SDK
codegen>=0.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

